+++ pyodide-build/pyodide_build/tests/test_common.py
@@ -1,9 +1,10 @@
-from pyodide_build.common import (
+from pyodide_build.common import (  # type: ignore[import]
     ALWAYS_PACKAGES,
     CORE_PACKAGES,
     CORE_SCIPY_PACKAGES,
     UNVENDORED_STDLIB_MODULES,
     _parse_package_subset,
+    find_matching_wheels,
     get_make_environment_vars,
     get_make_flag,
 )
@@ -77,3 +78,38 @@ def test_get_make_environment_vars():
     assert "SIDE_MODULE_CFLAGS" in vars
     assert "SIDE_MODULE_CXXFLAGS" in vars
     assert "TOOLSDIR" in vars
+
+
+def test_wheel_paths():
+    from pathlib import Path
+
+    old_version = "cp38"
+    PYMAJOR = int(get_make_flag("PYMAJOR"))
+    PYMINOR = int(get_make_flag("PYMINOR"))
+    current_version = f"cp{PYMAJOR}{PYMINOR}"
+    future_version = f"cp{PYMAJOR}{PYMINOR + 1}"
+    strings = []
+
+    for interp in [
+        old_version,
+        current_version,
+        future_version,
+        "py3",
+        "py2",
+        "py2.py3",
+    ]:
+        for abi in [interp, "abi3", "none"]:
+            for arch in ["emscripten_wasm32", "linux_x86_64", "any"]:
+                strings.append(f"wrapt-1.13.3-{interp}-{abi}-{arch}.whl")
+
+    paths = [Path(x) for x in strings]
+    assert [x.stem.split("-", 2)[-1] for x in find_matching_wheels(paths)] == [
+        f"{current_version}-{current_version}-emscripten_wasm32",
+        f"{current_version}-abi3-emscripten_wasm32",
+        f"{current_version}-none-emscripten_wasm32",
+        f"{old_version}-abi3-emscripten_wasm32",
+        "py3-none-emscripten_wasm32",
+        "py2.py3-none-emscripten_wasm32",
+        "py3-none-any",
+        "py2.py3-none-any",
+    ]