+++ src/core/pyproxy.ts
@@ -39,6 +39,10 @@ declare var IS_ITERATOR: number;
 declare var IS_AWAITABLE: number;
 declare var IS_BUFFER: number;
 
+declare var PYGEN_NEXT: number;
+declare var PYGEN_RETURN: number;
+declare var PYGEN_ERROR: number;
+
 declare function DEREF_U32(ptr: number, offset: number): number;
 // end-pyodide-skip
 
@@ -52,18 +56,23 @@ export function isPyProxy(jsobj: any): jsobj is PyProxy {
 }
 API.isPyProxy = isPyProxy;
 
+declare var FinalizationRegistry: any;
+declare var globalThis: any;
+
 if (globalThis.FinalizationRegistry) {
-  Module.finalizationRegistry = new FinalizationRegistry(([ptr, cache]) => {
-    cache.leaked = true;
-    pyproxy_decref_cache(cache);
-    try {
-      Module._Py_DecRef(ptr);
-    } catch (e) {
-      // I'm not really sure what happens if an error occurs inside of a
-      // finalizer...
-      API.fatal_error(e);
+  Module.finalizationRegistry = new FinalizationRegistry(
+    ([ptr, cache]: [ptr: number, cache: PyProxyCache]) => {
+      cache.leaked = true;
+      pyproxy_decref_cache(cache);
+      try {
+        Module._Py_DecRef(ptr);
+      } catch (e) {
+        // I'm not really sure what happens if an error occurs inside of a
+        // finalizer...
+        API.fatal_error(e);
+      }
     }
-  });
+  );
   // For some unclear reason this code screws up selenium FirefoxDriver. Works
   // fine in chrome and when I test it in browser. It seems to be sensitive to
   // changes that don't make a difference to the semantics.
@@ -155,8 +164,8 @@ Module.pyproxy_new = function (ptrobj: number, cache?: PyProxyCache) {
 };
 
 function _getPtr(jsobj: any) {
-  let ptr = jsobj.$$.ptr;
-  if (ptr === null) {
+  let ptr: number = jsobj.$$.ptr;
+  if (ptr === 0) {
     throw new Error(jsobj.$$.destroyed_msg);
   }
   return ptr;
@@ -236,7 +245,7 @@ function pyproxy_decref_cache(cache: PyProxyCache) {
 }
 
 Module.pyproxy_destroy = function (proxy: PyProxy, destroyed_msg: string) {
-  if (proxy.$$.ptr === null) {
+  if (proxy.$$.ptr === 0) {
     return;
   }
   let ptrobj = _getPtr(proxy);
@@ -254,7 +263,7 @@ Module.pyproxy_destroy = function (proxy: PyProxy, destroyed_msg: string) {
   // Maybe the destructor will call JavaScript code that will somehow try
   // to use this proxy. Mark it deleted before decrementing reference count
   // just in case!
-  proxy.$$.ptr = null;
+  proxy.$$.ptr = 0;
   destroyed_msg += "\n" + `The object was of type "${proxy_type}" and `;
   if (proxy_repr) {
     destroyed_msg += `had repr "${proxy_repr}"`;
@@ -311,14 +320,7 @@ Module.callPyObject = function (ptrobj: number, ...jsargs: any) {
   return Module.callPyObjectKwargs(ptrobj, ...jsargs, {});
 };
 
-export type Py2JsResult =
-  | PyProxy
-  | number
-  | bigint
-  | string
-  | boolean
-  | undefined;
-export type PyProxy = PyProxyClass & { [x: string]: Py2JsResult };
+export type PyProxy = PyProxyClass & { [x: string]: any };
 
 export class PyProxyClass {
   $$: { ptr: number; cache: PyProxyCache; destroyed_msg?: string };
@@ -578,7 +580,7 @@ export class PyProxyGetItemMethods {
    * @param key The key to look up.
    * @returns The corresponding value.
    */
-  get(key: any): Py2JsResult {
+  get(key: any): any {
     let ptrobj = _getPtr(this);
     let idkey = Hiwire.new_value(key);
     let idresult;
@@ -700,7 +702,7 @@ export class PyProxyContainsMethods {
  *
  * @private
  */
-function* iter_helper(iterptr: number, token: {}): Generator<Py2JsResult> {
+function* iter_helper(iterptr: number, token: {}): Generator<any> {
   try {
     let item;
     while ((item = Module.__pyproxy_iter_next(iterptr))) {
@@ -734,7 +736,7 @@ export class PyProxyIterableMethods {
    *
    * This will be used implicitly by ``for(let x of proxy){}``.
    */
-  [Symbol.iterator](): Iterator<Py2JsResult, Py2JsResult, Py2JsResult> {
+  [Symbol.iterator](): Iterator<any, any, any> {
     let ptrobj = _getPtr(this);
     let token = {};
     let iterptr;
@@ -780,27 +782,29 @@ export class PyProxyIteratorMethods {
    * some_value}``. When the generator raises a ``StopIteration(result_value)``
    * exception, ``next`` returns ``{done : true, value : result_value}``.
    */
-  next(arg: any = undefined): IteratorResult<Py2JsResult, Py2JsResult> {
-    let idresult;
+  next(arg: any = undefined): IteratorResult<any, any> {
     // Note: arg is optional, if arg is not supplied, it will be undefined
     // which gets converted to "Py_None". This is as intended.
     let idarg = Hiwire.new_value(arg);
+    let status;
     let done;
+    let stackTop = Module.stackSave();
+    let res_ptr = Module.stackAlloc(4);
     try {
-      idresult = Module.__pyproxyGen_Send(_getPtr(this), idarg);
-      done = idresult === 0;
-      if (done) {
-        idresult = Module.__pyproxyGen_FetchStopIterationValue();
-      }
+      status = Module.__pyproxyGen_Send(_getPtr(this), idarg, res_ptr);
     } catch (e) {
       API.fatal_error(e);
     } finally {
       Hiwire.decref(idarg);
     }
-    if (done && idresult === 0) {
+    let HEAPU32 = Module.HEAPU32;
+    let idresult = DEREF_U32(res_ptr, 0);
+    Module.stackRestore(stackTop);
+    if (status === PYGEN_ERROR) {
       Module._pythonexc2js();
     }
     let value = Hiwire.pop_value(idresult);
+    done = status === PYGEN_RETURN;
     return { done, value };
   }
 }
@@ -941,7 +945,7 @@ let PyProxyHandlers = {
     python_setattr(jsobj, jskey, jsval);
     return true;
   },
-  deleteProperty(jsobj: PyProxyClass, jskey: any) {
+  deleteProperty(jsobj: PyProxyClass, jskey: any): boolean {
     let descr = Object.getOwnPropertyDescriptor(jsobj, jskey);
     if (descr && !descr.writable) {
       throw new TypeError(`Cannot delete read only field '${jskey}'`);
@@ -955,7 +959,7 @@ let PyProxyHandlers = {
     python_delattr(jsobj, jskey);
     // Must return "false" if "jskey" is a nonconfigurable own property.
     // Otherwise JavaScript will throw a TypeError.
-    return !descr || descr.configurable;
+    return !descr || !!descr.configurable;
   },
   ownKeys(jsobj: PyProxyClass) {
     let ptrobj = _getPtr(jsobj);
@@ -977,7 +981,7 @@ let PyProxyHandlers = {
   },
 };
 
-export type PyProxyAwaitable = PyProxy & Promise<Py2JsResult>;
+export type PyProxyAwaitable = PyProxy & Promise<any>;
 
 /**
  * The Promise / JavaScript awaitable API.
@@ -1096,7 +1100,7 @@ export class PyProxyAwaitableMethods {
 
 export type PyProxyCallable = PyProxy &
   PyProxyCallableMethods &
-  ((...args: any[]) => Py2JsResult);
+  ((...args: any[]) => any);
 
 export class PyProxyCallableMethods {
   apply(jsthis: PyProxyClass, jsargs: any) {
@@ -1484,6 +1488,7 @@ export class PyBuffer {
       API.fatal_error(e);
     }
     this._released = true;
+    // @ts-ignore
     this.data = null;
   }
 }